backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
requests==2.32.3
pdfplumber==0.11.4
python-docx==1.1.2
nltk==3.9.1
gunicorn==22.0.0
